+++ packages/amazonq/test/e2e_new/amazonq/helpers/switchModelHelper.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import { By, WebviewView, WebElement } from 'vscode-extension-tester'
-import { waitForElement } from '../utils/generalHelper'
+import { waitForElement } from '../utils/generalUtils'
 
 /**
  * Lists all available model options in the dropdown
